Move open_storage out of __init__
